[NFC][SYCL] Run device-agnostic tests separately
Tests which do not require backend were run for every target backend. It
is waste of machine time. This PR:
 - move device-agnostic tests to separate target to run only once;
 - moved all device-specific LIT substitutions to on-device folder to
   avoid adding device-dependent tests to directory with
   device-agnostic tests;
 - move device-dependent tests to sycl/test/on-device folder.
