+++ sycl/test/CMakeLists.txt
@@ -21,6 +21,13 @@ configure_lit_site_cfg(
   )
 
 configure_lit_site_cfg(
+  ${CMAKE_CURRENT_SOURCE_DIR}/on-device/lit.site.cfg.py.in
+  ${CMAKE_CURRENT_BINARY_DIR}/on-device/lit.site.cfg.py
+  MAIN_CONFIG
+  ${CMAKE_CURRENT_SOURCE_DIR}/on-device/lit.cfg.py
+  )
+
+configure_lit_site_cfg(
   ${CMAKE_CURRENT_SOURCE_DIR}/Unit/lit.site.cfg.py.in
   ${CMAKE_CURRENT_BINARY_DIR}/Unit/lit.site.cfg.py
   MAIN_CONFIG
@@ -64,7 +71,7 @@ add_lit_testsuites(SYCL-DEPLOY ${CMAKE_CURRENT_SOURCE_DIR}
   )
 
 add_lit_testsuite(check-sycl-opencl "Running the SYCL regression tests for OpenCL"
-  ${CMAKE_CURRENT_BINARY_DIR}
+  ${CMAKE_CURRENT_BINARY_DIR}/on-device
   ARGS ${RT_TEST_ARGS}
   PARAMS "SYCL_PLUGIN=opencl"
   DEPENDS ${SYCL_TEST_DEPS}
@@ -73,21 +80,29 @@ add_lit_testsuite(check-sycl-opencl "Running the SYCL regression tests for OpenC
 set_target_properties(check-sycl-opencl PROPERTIES FOLDER "SYCL tests")
 
 add_lit_testsuite(check-sycl-level-zero "Running the SYCL regression tests for Level Zero"
-  ${CMAKE_CURRENT_BINARY_DIR}
+  ${CMAKE_CURRENT_BINARY_DIR}/on-device
   ARGS ${RT_TEST_ARGS}
   PARAMS "SYCL_PLUGIN=level_zero"
+  DEPENDS ${SYCL_TEST_DEPS}
+  EXCLUDE_FROM_CHECK_ALL
+  )
+
+add_lit_testsuite(check-sycl-spirv "Running device-agnostic SYCL regression tests for SPIR-V"
+  ${CMAKE_CURRENT_BINARY_DIR}
+  ARGS ${RT_TEST_ARGS}
+  PARAMS "SYCL_TRIPLE=spir64-unknown-linux-sycldevice"
   DEPENDS ${SYCL_TEST_DEPS}
   EXCLUDE_FROM_CHECK_ALL
   )
 set_target_properties(check-sycl-level-zero PROPERTIES FOLDER "SYCL tests")
 
 add_custom_target(check-sycl)
-add_dependencies(check-sycl check-sycl-opencl check-sycl-level-zero)
+add_dependencies(check-sycl check-sycl-opencl check-sycl-level-zero check-sycl-spirv)
 set_target_properties(check-sycl PROPERTIES FOLDER "SYCL tests")
 
 if(SYCL_BUILD_PI_CUDA)
   add_lit_testsuite(check-sycl-cuda "Running the SYCL regression tests for CUDA"
-    ${CMAKE_CURRENT_BINARY_DIR}
+    ${CMAKE_CURRENT_BINARY_DIR}/on-device
     ARGS ${RT_TEST_ARGS}
     PARAMS "SYCL_PLUGIN=cuda"
     DEPENDS ${SYCL_TEST_DEPS}
@@ -95,9 +110,17 @@ if(SYCL_BUILD_PI_CUDA)
   )
   set_target_properties(check-sycl-cuda PROPERTIES FOLDER "SYCL CUDA tests")
 
-  add_dependencies(check-sycl check-sycl-cuda)
+  add_lit_testsuite(check-sycl-ptx "Running device-agnostic SYCL regression tests for NVidia PTX"
+    ${CMAKE_CURRENT_BINARY_DIR}
+    ARGS ${RT_TEST_ARGS}
+    PARAMS "SYCL_TRIPLE=nvptx64-nvidia-cuda-sycldevice"
+    DEPENDS ${SYCL_TEST_DEPS}
+    EXCLUDE_FROM_CHECK_ALL
+  )
 
-  add_lit_testsuites(SYCL-CUDA ${CMAKE_CURRENT_SOURCE_DIR}
+  add_dependencies(check-sycl check-sycl-cuda check-sycl-ptx)
+
+  add_lit_testsuites(SYCL-CUDA ${CMAKE_CURRENT_SOURCE_DIR}/on-device
     PARAMS "SYCL_PLUGIN=cuda"
     DEPENDS ${SYCL_TEST_DEPS}
     EXCLUDE_FROM_CHECK_ALL
