+++ sycl/test/on-device/config/device_selector.cpp
